test: All mocks now belong to their own packages
