+++ packages/core-ethereum/src/index.ts
@@ -22,23 +22,24 @@ import { createChainWrapper } from './ethereum'
 import { PROVIDER_CACHE_TTL } from './constants'
 import { EventEmitter } from 'events'
 import { initializeCommitment, findCommitmentPreImage, bumpCommitment } from './commitment'
+import { chainMock } from './index.mock'
 
 const log = debug('hopr-core-ethereum')
 
 export type RedeemTicketResponse =
   | {
-      status: 'SUCCESS'
-      receipt: string
-      ackTicket: AcknowledgedTicket
-    }
+    status: 'SUCCESS'
+    receipt: string
+    ackTicket: AcknowledgedTicket
+  }
   | {
-      status: 'FAILURE'
-      message: string
-    }
+    status: 'FAILURE'
+    message: string
+  }
   | {
-      status: 'ERROR'
-      error: Error | string
-    }
+    status: 'ERROR'
+    error: Error | string
+  }
 
 export default class HoprEthereum extends EventEmitter {
   public indexer: Indexer
@@ -296,7 +297,7 @@ export default class HoprEthereum extends EventEmitter {
     let c: ChannelEntry
     try {
       c = await this.db.getChannelTo(dest)
-    } catch {}
+    } catch { }
     if (c && c.status !== ChannelStatus.Closed) {
       throw Error('Channel is already opened')
     }
@@ -324,6 +325,7 @@ export default class HoprEthereum extends EventEmitter {
 export {
   ChannelEntry,
   Indexer,
+  chainMock,
   createChainWrapper,
   initializeCommitment,
   findCommitmentPreImage,
