+++ packages/core/src/index.mock.ts
@@ -1,3 +1,3 @@
-import { HoprOptions } from '@hoprnet/hopr-core'
+import { HoprOptions } from '.'
 
 export const sampleOptions = { environment: { id: '1' } } as unknown as HoprOptions
