+++ packages/core/src/libp2p.mock.ts
@@ -2,10 +2,16 @@ import LibP2P from 'libp2p'
 import ConnectionManager from 'libp2p/src/connection-manager'
 import PeerStore from 'libp2p/src/peer-store'
 import AddressManager from 'libp2p/src/address-manager'
-import { debug } from '@hoprnet/hopr-utils'
-import { NAMESPACE, mockPeerId, sampleMultiaddrs, samplePeerId } from './constants'
+import { debug, privKeyToPeerId } from '@hoprnet/hopr-utils'
+import PeerId from 'peer-id'
+import { Multiaddr } from 'multiaddr'
 
-const libp2pLogger = debug(`${NAMESPACE}:libp2p`)
+export const privateKey = '0xcb1e5d91d46eb54a477a7eefec9c87a1575e3e5384d38f990f19c09aa8ddd332'
+export const mockPeerId = privKeyToPeerId(privateKey)
+export const samplePeerId = PeerId.createFromB58String('16Uiu2HAmThyWP5YWutPmYk9yUZ48ryWyZ7Cf6pMTQduvHUS9sGE7')
+export const sampleMultiaddrs = new Multiaddr(`/ip4/127.0.0.1/tcp/124/p2p/${samplePeerId.toB58String()}`)
+
+const libp2pLogger = debug(`hopr:mocks:libp2p`)
 let libp2p: LibP2P
 
 libp2p = {} as unknown as LibP2P
