+++ packages/utils/src/index.ts
@@ -11,5 +11,6 @@ export * from './parseJSON'
 export * from './time'
 export * from './constants'
 export * from './db'
+export * from './db.mock'
 export * from './ethereum'
 export * from './utils'
